Removing monkey patching for Characterization job
The previous commit represented the separation of how Sufia assumed
file uploads with virus check behaved versus how we needed them to
work. The effort is to consolidate those into a single file.
